chore: rename quic.[ch] -> Quic.[ch]
