+++ src/CMakeLists.txt
@@ -334,8 +334,8 @@ IF (PAHO_WITH_SSL)
 
     TARGET_LINK_LIBRARIES(paho-mqtt3cs PRIVATE msquic)
     TARGET_LINK_LIBRARIES(paho-mqtt3as PRIVATE msquic)
-    TARGET_SOURCES(paho-mqtt3as PRIVATE quic.c)
-    TARGET_SOURCES(paho-mqtt3cs PRIVATE quic.c)
+    TARGET_SOURCES(paho-mqtt3as PRIVATE Quic.c)
+    TARGET_SOURCES(paho-mqtt3cs PRIVATE Quic.c)
     SET_PROPERTY(TARGET common_ssl_obj PROPERTY COMPILE_DEFINITIONS "MSQUIC=1;OPENSSL=1;PAHO_MQTT_EXPORTS=1")
     TARGET_COMPILE_OPTIONS(common_ssl_obj PUBLIC "-ggdb3")
     SET_PROPERTY(TARGET paho-mqtt3as PROPERTY COMPILE_DEFINITIONS "MSQUIC=1;OPENSSL=1;PAHO_MQTT_EXPORTS=1")
